cosmwasm: Extract CosmosMsg types into separate file
