+++ packages/cosmwasm/src/cw3cosmwasmclient.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { BroadcastMode, GasLimits, GasPrice, OfflineSigner } from "@cosmjs/launchpad";
 
+import { CosmosMsg } from "./cosmosmsg";
 import { Account } from "./cosmwasmclient";
-import { CosmosMsg } from "./msgs";
 import { CosmWasmFeeTable, ExecuteResult, SigningCosmWasmClient } from "./signingcosmwasmclient";
 
 export type Expiration =
